Regressionrefactor

+++ cucumber/src/main/java/com/jcrew/steps/StartingSteps.java
@@ -17,7 +17,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.concurrent.TimeUnit;
 
 public class StartingSteps {
 
@@ -30,8 +29,15 @@ public class StartingSteps {
 
     @Before
     public void setupDriver() throws IOException {
+    	stateHolder.put("deletecookies", false);
         driverFactory = new DriverFactory();
         driver = driverFactory.getDriver();
+    }
+
+    @Given("User is on homepage with clean session")
+    public void user_is_on_home_page_with_clean_session() {
+        driverFactory.deleteBrowserCookies();
+        user_is_on_home_page();
     }
 
     @Given("^User is on homepage$")
@@ -112,19 +118,29 @@ public class StartingSteps {
         Util.waitForPageFullyLoaded(driver);
     }
 
+    @And("^Deletes browser cookies$")
+    public void deletes_browser_cookies(){
+        driverFactory.deleteBrowserCookies();
+        stateHolder.put("deletecookies", true);
+    }
+
     @After
     public void quitDriver(Scenario scenario) throws IOException {
 
         if (driver != null && (scenario.isFailed() || scenario.getName().contains(TAKE_SCREENSHOT))) {
-
             logger.debug("Taking screenshot of scenario {}", scenario.getName());
-
-            final byte[] screenshot = ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
-            scenario.embed(screenshot, "image/png");
+            try {
+                final byte[] screenshot = ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
+                scenario.embed(screenshot, "image/png");
+                deletes_browser_cookies();
+            } catch (Exception e){
+                logger.error("An exception happened when taking step screenshot after scenario", e);
+                driverFactory.resetDriver();
+            }
         }
 
-        if (driverFactory != null) {
-           driverFactory.destroyDriver();
+        if (driverFactory != null && (boolean)stateHolder.get("deletecookies")) {
+            driverFactory.destroyDriver();
         }
 
         stateHolder.clear();
@@ -145,7 +161,7 @@ public class StartingSteps {
                 }
 
             } catch (Exception e) {
-                logger.error("An exception happened when taking step screenshot", e);
+                logger.error("An exception happened when taking step screenshot after step", e);
             }
         }
     }
